refactor: environment variable handling and remove unnecessary setup file
