--- jest.setup.ts
@@ -1,3 +1,0 @@
-import dotenv from 'dotenv';
-
-dotenv.config();
