+++ package.json
@@ -4,9 +4,9 @@
   "description": "An Express.js boilerplate backend",
   "main": "index.ts",
   "scripts": {
-    "dev": "tsx watch --clear-screen=false -r dotenv/config src/index.ts",
+    "dev": "tsx watch --clear-screen=false src/index.ts",
     "build": "rimraf build && tsc",
-    "start": "npm run build && NODE_ENV=production tsx -r dotenv/config build/index.js",
+    "start": "npm run build && NODE_ENV=production tsx build/index.js",
     "lint": "eslint 'src/**/*.{ts,js}'",
     "lint:fix": "eslint --fix --ext .ts,.js src/",
     "format": "prettier --write 'src/**/*.{ts,js,json}'",
@@ -61,9 +61,6 @@
   "jest": {
     "preset": "ts-jest",
     "testEnvironment": "node",
-    "setupFilesAfterEnv": [
-      "<rootDir>/jest.setup.ts"
-    ],
     "roots": [
       "<rootDir>/src"
     ],
