+++ src/common/utils/envConfig.ts
@@ -1,10 +1,3 @@
-import * as dotenv from 'dotenv';
-import path from 'path';
-
-dotenv.config({
-  path: path.resolve(__dirname, '../../.env'),
-});
-
 export const getPort = () => getEnvVar<number>('PORT', 'number');
 export const getNodeEnv = () => getEnvVar<string>('NODE_ENV', 'string');
 export const getCorsOrigin = () => getEnvVar<string>('CORS_ORIGIN', 'string');
